+++ godot-scene-signal-auditor/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 0.1.1 - 2026-06-09
+
+- Ensured file output reports end with a trailing newline.
+
 ## 0.1.0 - 2026-06-08
 
 - Added `.tscn` persistent connection parser.